add WT modules
