+++ .gitignore
@@ -1,2 +1,3 @@
 **/Packages/
 **/_PackageInt/
+**/node_modules/
