+++ PackageSources/html_ui/Pages/VCockpit/Instruments/Airliners/CS777-200LR/FMC/CS777-200LR_FMC.html
@@ -7,20 +7,20 @@
 </script>
 
 <script type="text/html" import-script="/JS/SimPlane.js"></script>
-<script type="text/html" import-script="/Pages/Salty/WT/WT_SimPlaneMod.js"></script> <!-- WT -->
+<script type="text/html" import-script="/Pages/RobSim/WT/WT_SimPlaneMod.js"></script> <!-- WT -->
 <script type="text/html" import-script="/JS/dataStorage.js"></script>
-<script type="text/html" import-script="/Pages/Salty/WT/DataStore.js"></script> <!-- WT -->
-<script type="text/html" import-script="/Pages/Salty/WT/WTUtils.js"></script> <!-- WT -->
-<script type="text/html" import-script="/Pages/Salty/WT/WT_ConvertUnit.js"></script> <!-- WT -->
-<script type="text/html" import-script="/Pages/Salty/WT/WT_GCMath.js"></script> <!-- WT -->
+<script type="text/html" import-script="/Pages/RobSim/WT/DataStore.js"></script> <!-- WT -->
+<script type="text/html" import-script="/Pages/RobSim/WT/WTUtils.js"></script> <!-- WT -->
+<script type="text/html" import-script="/Pages/RobSim/WT/WT_ConvertUnit.js"></script> <!-- WT -->
+<script type="text/html" import-script="/Pages/RobSim/WT/WT_GCMath.js"></script> <!-- WT -->
 
 <script type="text/html" import-script="/Pages/VCockpit/Instruments/Shared/Utils/RadioNav.js"></script>
 
 <script type="text/html" import-script="/Pages/VCockpit/Instruments/Shared/FlightElements/Runway.js"></script>
-<script type="text/html" import-script="/Pages/Salty/WT/Waypoint.js"></script> <!-- WT -->
+<script type="text/html" import-script="/Pages/RobSim/WT/Waypoint.js"></script> <!-- WT -->
 <script type="text/html" import-script="/Pages/VCockpit/Instruments/Shared/FlightElements/Approach.js"></script>
 <script type="text/html" import-script="/Pages/VCockpit/Instruments/Shared/FlightElements/FlightPlan.js"></script>
-<script type="text/html" import-script="/Pages/Salty/WT/WaypointLoader.js"></script> <!-- WT -->
+<script type="text/html" import-script="/Pages/RobSim/WT/WaypointLoader.js"></script> <!-- WT -->
 <script type="text/html" import-script="/Pages/Salty/fpm/LzUtf8.js"></script> <!-- WT -->
 <script type="text/html" import-script="/Pages/Salty/fpm/wtsdk.js"></script> <!-- WT -->
 <script type="text/html" import-script="/Pages/VCockpit/Instruments/Shared/FlightElements/GeoCalc.js"></script>
@@ -46,14 +46,14 @@
 
 
 <!-- WT -->
-<script type="text/html" import-script="/Pages/Salty/WT/Autopilot/NavToNavTransfer.js"></script>
-<script type="text/html" import-script="/Pages/Salty/WT/Autopilot/SaltyNavModeSelector.js"></script>
-<script type="text/html" import-script="/Pages/Salty/WT/Autopilot/WT_BaseVnav.js"></script>
-<script type="text/html" import-script="/Pages/Salty/WT/Autopilot/AutopilotMath.js"></script>
-<script type="text/html" import-script="/Pages/Salty/WT/Autopilot/HoldsDirector.js"></script>
-<script type="text/html" import-script="/Pages/Salty/WT/Autopilot/LocDirector.js"></script>
-<script type="text/html" import-script="/Pages/Salty/WT/Autopilot/LNavDirector.js"></script>
-<script type="text/html" import-script="/Pages/Salty/WT/Autopilot/WT_VnavAutopilot.js"></script>
+<script type="text/html" import-script="/Pages/RobSim/WT/Autopilot/NavToNavTransfer.js"></script>
+<script type="text/html" import-script="/Pages/RobSim/WT/Autopilot/SaltyNavModeSelector.js"></script>
+<script type="text/html" import-script="/Pages/RobSim/WT/Autopilot/WT_BaseVnav.js"></script>
+<script type="text/html" import-script="/Pages/RobSim/WT/Autopilot/AutopilotMath.js"></script>
+<script type="text/html" import-script="/Pages/RobSim/WT/Autopilot/HoldsDirector.js"></script>
+<script type="text/html" import-script="/Pages/RobSim/WT/Autopilot/LocDirector.js"></script>
+<script type="text/html" import-script="/Pages/RobSim/WT/Autopilot/LNavDirector.js"></script>
+<script type="text/html" import-script="/Pages/RobSim/WT/Autopilot/WT_VnavAutopilot.js"></script>
 
 <!-- FMC -->
 <!-- FMC B777 Profile -->
