+++ PackageSources/manifest.json
@@ -17,4 +17,5 @@
       "OlderHistory": ""
     }
   },
-  "total_package_size": "00000000000137855371"
+  "total_package_size": "00000000000138311614"
+}
