+++ package.json
@@ -24,5 +24,8 @@
   "bugs": {
     "url": "https://github.com/RobSimulation/robsim-salty-777/issues"
   },
-  "homepage": "https://github.com/RobSimulation/robsim-salty-777#readme"
+  "homepage": "https://github.com/RobSimulation/robsim-salty-777#readme",
+  "dependencies": {
+    "fs-extra": "^10.0.0"
+  }
 }
